net: xilinx_emacps: Fix MII initialization
MII probing should be done in open function.

The problem was that phydev was connected in the xemacps_probe
function and disconnected in the xemacps_close function.

It means that if device is closed and open again phydev
is not connected because phydev connection is not setup.
It ends with NULL pointer exception because
phydev->attached_dev(in phy_detach()) is NULL.

The bug was introduced by this commit:
"net: xilinx_emacps: Synchronization with PetaLinux tree"
(sha1: 8222a810a56b10b67be3d75fe3ac25150e19c8c2)

Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Tested-by: John Linn <john.linn@xilinx.com>
Tested-by: Soren Brinkmann <soren.brinkmann@xilinx.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>