xilinx_emacps: Dont mutex the rx and tx paths on ISR
The if-else logic of the ISR is such that only on of the RX or TX paths
are scheduled for action on interrupt (preference to TX). If the interrupts
for TX and RX occur concurrently, then TX will happen RX wont, but both will
get their interrupt status cleared. Fixed by removing the if-else.

Signed-off-by: Peter Crosthwaite <peter.crosthwaite@xilinx.com>
Acked-by: Anirudha Sarangi <anirudha.sarangi@xilinx.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>