xilinx_emacps: Fix for Rx lockdown HW issue
Under heavy traffic of small UDP packets, the Rx path in GEM
locks up. This is because of a HW bug w.r.t handling of resource errors.
Under a heavy resource errors scenario one of the internal counters in
the HW misbehaves resulting in a overflow condition and reducing the
availabe Rx SRAM. This patch provides a SW workaround for the HW bug.
The SW workaround involves first trying to avoid hitting the issue and
if eventually we hit it, we come out of the issue gracefully.
For every resource error, we set the bit 18 in the net_ctrl register
to flush a packet out of Rx SRAM. This avoids hitting the issue.
The SW workaround also uses the existing general purpose timer (100 msec)
(used to update the stats) and reads the received frame counter stat
register on each call to timer handler. If in two consecutive
timer handler calls, the counter do not increment, the SW toggles
the rx_enable bit to do a SW reset of the Rx logic.

Signed-off-by: Anirudha Sarangi <anirudh@xilinx.com>
Reviewed-by: Peter Crosthwaite <peter.crosthwaite@xilinx.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>