qbman: don't initialize the qbman again in kexec boot
We can't do a hardware reset for qbman. So it may causes unknown
issues when changing the memories used by them at runtime.
Previously we have used this method for the kdump kernel, and it
should be safe for kexec boot too. Also add a WRAN_ON if we
can't reserve the memories successfully.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Bin Jiang <bin.jiang@windriver.com>