bman: drain all the buffer pools before kdump boot
In the kdump boot, there are still buffers in the bman buffer pools
released by the first kernel. These buffers have nothing to do with
the current kernel, and may cause issues when we try to use these
buffers. So we should drain all the buffer pools before kdump boot.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Bin Jiang <bin.jiang@windriver.com>