net: xilinx_emacps: Support half-duplex mode in xemacps_adjust_link().
In xemacps_adjust_link(), pass half-duplex modes to the gmii_to_rgmii
core. For full-duplex links, the behaviour is the same as before (the
duplex bit was always set).

Also, merge the two sets of "if" blocks (where the speed was checked
twice), and add a case for unexpected PHY speeds.

Signed-off-by: Thomas Betker <thomas.betker@rohde-schwarz.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>