net: xilinx_emacps: Add PTP hardware clock support
This patch adds timestamping and PTP hardware clock support to the
emacps driver.
Removed the existing timestamp implementation as the support for this
implementation is deprecated in kernel and causing the compilation errors.

With new implementation, Hardware frequency adjustment algorithm is not working
properly and this will be fixed in future versions.

Signed-off-by: Punnaiah Choudary Kalluri <punnaia@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>