remoteproc/pruss: define platform data for PRUSS bus driver
This commit comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

The PRUSS can have a PRCM reset line associated with the IP on
some OMAP architecture based SoCs. The reset needs to be programmed
properly before accessing any of the internal registers in the PRUSS.
This functionality is achieved through the omap_device layer, which
is not exposed outside of mach-omap2 layer. Define a platform data
structure for PRUSS so that this API can be invoked through platform
data ops from the driver.

Signed-off-by: Suman Anna <s-anna@ti.com>
(cherry picked from commit 883af303c21b5f00b4fec96d475181b7bbfabd30)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>