net: xemacps: Make sure device is active when MDIO bus is accessed
Due to runtime PM it could happen that the GEM was runtime suspended,
having its clock disabled, but still MDIO access could happen. Those
require the GEM registers to be accessible, which requires the clocks to
be enabled.
Add runtime_pm_(get|put)() calls in the MDIO access path to ensure the
GEM clocks are enabled when the MDIO bus is accesses.

Signed-off-by: Soren Brinkmann <soren.brinkmann@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>