ARM: zynq: Unconditionally shutdown ARMPLL during suspend
This moves the ARMPLL shutdown from the early suspend stage to very
late.
Originally the bypass/shutdown was tried to be graceful using the CCF.
But that resulted in problems with Zynq timers that all depend on the
CPU clock.
The new approach moves the shutdown to very late hoping to avoid all the
timer related troubles, but also implies some - now mandatory -
assumptions:
 - peripherals - especially the wake up source - are clocked by the
   IOPLL
 - Nothing else in the chip relies on the ARMPLL to stay stable

Signed-off-by: Soren Brinkmann <soren.brinkmann@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>