drm: xilinx: crtc/plane: Decouple the plane manager functions
With the universal plane feature, the CRTC and the plane can be
configured independently. This patch decouples the plane and
the plane manager functions, so that the plane can be configured
separately from the CRTC. The PLANE driver configures planes
with plane functions, while the CRTC can use plane manager
functions along with plane functions.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>