net: emacps: Avoid Rx engine reset for 10 and 100mbps speeds
The Rx engine lock down issue, occurs when there is a heavy rx traffic
of smaller packets, is observed when the controller is operating at
1000 mbps speeds. But the driver workaround for this issue is added
for all the speeds. The recent series of tests (in house and customer level)
shows that the lock down issues was not observed for 10 and 100 speeds and
the existing workaround causes unpredictable results. so, restring the
workaround to 1000mbps speed only.

Signed-off-by: Punnaiah Choudary Kalluri <punnaia@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>