drm: xilinx: fb: Use xilinx_drm_format_bpp() as fallback
drm_fb_get_bpp_depth() only supports RGB formats, and returns
depth and bits per pixel values as '0' for YUV formats. In that case,
use xilinx_drm_format_bpp() to get the 'bits per pixel' value for
YUV formats, as the value is required to cacluate the size of framebuffer
and DMA transaction.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Reported-by: Christian Kohn <christian.kohn@xilinx.com>
Tested-by: Christian Kohn <christian.kohn@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>