spi: zynq-qspi: Use 0xff to pad Tx data
When padding Tx data, the pad byte is currently undefined; set it
to 0xff in zynq_qspi_copy_write_data().

Tx data padding is used to align the data sent with a Program
command to a multiple of two bytes. Sending zero bits may overwrite a
flash byte which is not intended to be written; send 0xff to keep the
extra flash byte unchanged.

Signed-off-by: Thomas Betker <thomas.betker@rohde-schwarz.com>
Reviewed-by: Harini Katakam <harinik@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>