spi: zynq-qspi: Revise copy_read_data
Combine zynq_qspi_copy_read_data() with reading the RxFIFO register, in
analogy to the Tx path; rename the function to zynq_qspi_read_rx_fifo().

The new function reads the RXD register and copies the data to
xqspi->rxbuf. It also handles the pad byte in the dual-parallel case.

This is just a cleanup; the control flow is not affected

Signed-off-by: Thomas Betker <thomas.betker@rohde-schwarz.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Reviewed-by: Harini Katakam <harinik@xilinx.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>