spi: zynq-qspi: Add drain_rx_fifo function
Factor out zynq_qspi_drain_rx_fifo() from zynq_qspi_irq(), for
readability.

Rewrite the loop to make clear that we read min(len, 4*rxcount) bytes
from the RxFIFO, which is what we want. The old code does the same,
provided that len%4==0 || txlen==0; this condition is satisfied due to
the FIFO fill/drain pattern. [When the condition is not satisfied, the
old code reads more than the desired number of bytes; all this is not
readily apparent, though.]

Signed-off-by: Thomas Betker <thomas.betker@rohde-schwarz.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>