net: phy: Use proper duplex settings for SGMII case
The duplex setting is not set by the Xilinx phy driver
Which causes issues such that SGMII won't switch rates.
When the duplex is not set it is a value of -1 which then messes
Up the adjust_link() function in the driver as -1 is 0xFFFFFFFF
Which than causes it not to see any link state change.

Reported-by: John Linn <john.linn@xilinx.com>
Signed-off-by: Kedareswara rao Appana <appanad@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>