mmc: sdhci-of-esdhc: add eMMC HS200 mode support
The eSDHC is not the standard SD host controller of SD spec. To
support eMMC HS200 mode, a Freescale eMMC4.5 adapter card will be
used for hardware support and the peripheral clock will be used
instead of the platform clock. The driver should also define an
esdhc-specific esdhc_set_tuning_block() function for its tuning
process.

Signed-off-by: Yangbo Lu <yangbo.lu@nxp.com>
[Original patch taken from QorIQ-SDK-V2.0-20160527-yocto]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>