drm: xilinx: dp_sub: Set global alpha and alpha for single layer
Even when only one layer is enabled, the blender blends the active layer
with 0x0 value of the current format of the inactive layer. To work
around this confusion, driver implicitly enables the global alpha
and sets the alpha value to make only one layer is visible. When
both layers are enbled, values are reverted to requested values.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Reviewed-by: Radhey Shyam Pandey <radheys@xilinx.com>
Tested-by: Radhey Shyam Pandey <radheys@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>