fmd: Fix FMan engine pointer for FM_VSP_Config
In compatibility mode (32bit applications working with 64bit kernel),
when calling the FM_VSP_Config function there is a problem when
assigning the pointer to the FMan engine control block. Basically,
the application will send the VSP configuration call to the proper
FMan engine device file (let it be FMan #0 or FMan #1 device file).
The address to this engine is not available in user space hence
translating it from the application parameters is useless. The only
thing the application CAN send is the NUMBER of the FMan engine, but
never its address. This address should basically be recovered directly
from the device file the FM_VSP_Config call came from, in the same
way it is done for non-compatibility mode (32bit applications working
with 32bit kernel or 64bit applications working with 64bit kernel).

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>
[Original patch taken from QorIQ-SDK-V2.0-20160527-yocto]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>