xhci: iommu: add stream ID configuration for USB ports
commit  2329627a1302ea7b1e3c1c2bcb7e73e50a971efb from
https://github.com/MarvellEmbeddedProcessors/linux-marvell.git

A stream ID is a unique number that identifies the device
when it requires translation on the SMMU (MMU-500).
The stream IDs for the units in CP-110 are configured in firmware
and are added to DT to be used by the SMMU driver.

Change-Id: Ifd87535a936b680db733e0e80ff6c752e2179cac
Signed-off-by: Yehuda Yitschak <yehuday@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/29161
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Tested-by: Star_Automation <star@marvell.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>