mtd: spi-nor: fsl-quad: Add flash S25FS extra support
There are some boards have the same QSPI controller but have
different vendor falsh, So as to controller can use the same
compatible and share the driver, Just for different flash to do
the appropriate adaptation. Based on this, we need add the vendor
field in spi-nor, Because we will use the field to distribute
corresponding LUT for different flash operations.

Signed-off-by: Yunhui Cui <yunhui.cui@nxp.com>
Signed-off-by: Yuan Yao <yao.yuan@nxp.com>
[Original patch taken from NXP LSDK1709.]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>