net: ethernet: xilinx: Fix race condition for 10G MAC
In the driver napi is enabled after interrupts
are being registered. This may casue issue
if there is already data in the mac rx fifo.

If there is already data in the mac rx fifo
in the current driver flow it will trigger RX ISR
there we are disabling RX interrupts again it will
get enabled after processing the packets upto quota
in the napi poll.

But with the current driver flow napi will never be called
resulting no more rx packets being processed(As napi is enabled
at the end of the open function).

This patch fixes this issue.

Signed-off-by: Kedareswara rao Appana <appanad@xilinx.com>
Acked-by: Harini Katakam <harinik@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>