net: dpaa_eth: simplify the .shutdown() function of fman driver
The .shutdown() function is only invoked when reboot or kexec boot.
Making it the same as .remove() seems a bit overkill. Especially in
kexec boot, the .shutdown() of fman driver may be invoked before the
one of dpa driver due to the order of fman and dpa device node changed
in dtb. But the .shutdown() of dpa driver may depend on the existing of
some structs in fm driver, so we implement a simplified version of
.shutdown() only to do the necessary cleanup to workaround this issue.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>