drm: xilinx: drv: Set all clients as master
commit  b0e1a5fd84d0f16dea15229c5827142825b3b986 from
https://github.com/Xilinx/linux-xlnx.git

With this change, any client with root privilege will be master.
This means that all applications that run on the Xilinx DRM driver
as root get full control of the driver. It is hacky, but easy way
to enable multiple clients, for example including X11 and
other KMS applications to control of the pipeline.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>