MLK-13361-1 arm: imx6q: busfreq: rename 100MHz-related macros
Two macros are renamed:

1) set_timings_above_100MHz_operation as restore_mmdc_settings_info
2) mmdc_clk_lower_100MHz as mmdc_clk_lower_equal_100MHz

For (1) the operation is generic to several cases and not just related
(at least on a semantic way) with the operations "above" 100MHz

Renamed as restore_mmdc_settings_info the  macro can be reused for the
other cases like equal to 100MHz and possibly other intermediate
operation points.

For (2), the macro is renamed as mmdc_clk_lower_equal_100MHz to reflect
that this macro handles both the "lower than 100 MHz" case and the
"equal to 100MHz" case.

Signed-off-by: Juan Gutierrez <juan.gutierrez@nxp.com>
[Quanyang:Original patch taken from
https://source.codeaurora.org/external/imx/linux-imx.git imx_4.9.88_2.0.0_ga]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>