dts: a80x0: a70x0: a37x0: add customer board example device-tree
commit  5ff61c0e2e050bf415c4db03389f6b0bd155a5e4 from
https://github.com/MarvellEmbeddedProcessors/linux-marvell.git

This patch adds the following device-trees:
	armada-8040-customer.dts
	armada-7040-customer.dts
	armada-3700-customer.dts

Those device-trees should be used only as an example for
a customer porting.
All interfaces are disabled, so using those device-trees is
not possible.

A costumer needs to modify the nodes according to the board
which is used, for example:
	- model: "Costumer board model"
	- compatible: "costumer-board-compatible"
	- MPPs: pinctrl-0 = <handle>
	- PHY addresses for network nodes
	- COMPHY handles for network nodes
	- SPI flash (if needed)
	- I2c devices (if needed)
	- Enable nodes

For more information refer to: ./Documentation/mvebu/porting-guide.txt

Change-Id: I27edc83d71c9a4333a4444312602498e714a55f5
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/34011
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: Omri Itach <omrii@marvell.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>