ARM64: zynqmp: Remove rollback in pm domain power off in case of error
commit  255cc5550dc1895b0b8d5e35f6edb5e177980967 from
https://github.com/Xilinx/linux-xlnx.git

Removed reversing the power off operation in the case of an error.
An error could appear only if something is wrong with the communication
to the PFW and in that case, reversing the operation makes no sense
because it relies on the same communication channel. Added print for an
error case.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-by: Soren Brinkmann <soren.brinkmann@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>