ARM64: zynqmp: Added request/release node for pm domains
commit  5840e19729f7c654e84c37664a78e912e00d73be from
https://github.com/Xilinx/linux-xlnx.git

PM nodes must be requested before the pm domain is used, and released
after the domain becomes unused. That way, the PFW grants the
permissions to power on/off a domain. The request node API is called
for all PM nodes mapped to the pm domain when the first device is
attached to the domain. This happens upon boot and a driver bind.
The release node is called for all PM nodes mapped to the domain
when the last device is detached from the domain. A device is
detached on driver unbind.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Signed-off-by: Filip Drazic <filip.drazic@aggios.com>
Acked-by: Will Wong <willw@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>