ARM: dts: am33xx: Reserve pm code and data regions in ocmcram sram node
Add a 'pm_sram_code' reserved region to the ocmcram node to be exposed
by the mmio-sram driver as a pool but also mark it protect-exec so that
it can run code copied to it using sram_exec_copy.

Add another 'pm_sram_data' reserved region to the ocmcram node to act as
the data space for any code running from the 'pm_sram_code' region that
is exposed as a regular pool.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
(cherry picked from commit 842fe494a156f00ea838bae667c50fb3edf39758)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>