wkup_m3_ipc: Add support for IO Isolation
This commit comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

AM43xx support isolation of the IOs so that control is taken
from the peripheral they are connected to and overridden by values
present in the CTRL_CONF_* registers for the pad in the control module.

The actual toggling happens from the wkup_m3, so use a DT property from
thea wkup_m3_ipc node to allow the PM code to communicate the necessity
for placing the IOs into isolation to the firmware.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
(cherry picked from commit 89082747f9bfaba58d7f21dce30d20779d50f6b7)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>