OMAP: CLK: CLKSRC: Add suspend resume hooks
This commit comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

Add the save and restore for clksrc as part of suspend and resume
so that it saves the counter value and restores. This is needed in
modes like rtc+ddr in self-refresh not doing this stalls the time.

Signed-off-by: Keerthy <j-keerthy@ti.com>
(cherry picked from commit 831e06e05a709e178c36268c6b80b3731b7e6051)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>