fsl_usdpaa.c: Deprecate undefined dma_mem_create()
Deprecate undefined semantics when accessing shared memory.  The size
parameter in dma_mem_create() is ambiguous in the case of getting a
handle to an existing shared memory. Does the original size of the map
apply? Does the new size "grow/shrink" the memory map?  Internally
this case was not handled. A caller was able to get a handle to a
shared memory even if an incorrect size was specified.  If the user
then tries to access memory outside of the original shared map they
got a segfault. This commit removes this ambiguity. The caller is
still allowed to specify a size. Length must either match the existing
memory map size or it must be 0. The returned mapping will have the
correct size in both cases.

JIRA issue#: QLINUX-6519

Signed-off-by: Ahmed Mansour <ahmed.mansour@nxp.com>
[Original patch taken from SDK-V2.0-1703]
Signed-off-by: Meng Li <Meng.Li@windriver.com>