MXSCM-243-2 arm: imx6q: mmdc handshake for lpddr2 2ch-mode on low-power
For i.mx6q systems the mmdc handshake on channel 0 is kept enabled (while
channel 1 is bypassed). This is ok for lpddr2 systems operating on 1ch-mode,
but not true for 2ch-mode. On this case the handshake needs to be set for
both channels, otherwise a kernel panic or Oops error might be observed
after resuming from suspend.

Signed-off-by: Juan Gutierrez <juan.gutierrez@nxp.com>
[Quanyang:Original patch taken from
https://source.codeaurora.org/external/imx/linux-imx.git imx_4.9.88_2.0.0_ga]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>