arm64: dts: marvell: add interrupt description for pca io-expander
commit  d8e007315a9c8589fa6848451933de6b523d4c18 from
https://github.com/MarvellEmbeddedProcessors/linux-marvell.git

The MPP62 have to be configured as an GPIO and it is used as an interrupt
source for the whole io-expander and not only for u3d vbus-gpio. Therefore
remove wrong description of u3d_vbus_pins, it references and use pca0_pins
as pinctrl of io-expander.

Change-Id: I8d5d47c764ab98a1584b82b56c3926b4cd487fd5
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/38541
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>