remoteproc: zynqmp_r5: Allow IPI interrupt when RPU is not started
commit  2f13cf0fad05d63711b39b49dc88fb9464f65d8a from
https://github.com/Xilinx/linux-xlnx.git

It is possible that when the remote is running when the remoteproc
driver is probed.
There can be notification between the Linux and the remote on vdev
resource entry changes without starting the remote.

This patch enables IPI interrupt in local init().
It will make sure IPI is enabled in start() as it is possible to
have stop() before start().

Signed-off-by: Wendy Liang <jliang@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>