remoteproc: zynq: use sram for firmware memory
commit  8973dece8eb2d982d4c8b6c444789759fd961fd5 from
https://github.com/Xilinx/linux-xlnx.git

Previously, declare some DMA memory and use carveout memory
for the firmware. However, allocating DMA memory for the
firmware cannot guarantee the memory allocated is the expected
memory defined in the firmware.

Based on experiment, the base address of the allocated DMA
memory can vary due to the firmware size.

We should follow the other remoteproc driver, if we want the
exact physical address for the firmware, we should da_to_va()
op defined by the remoteproc driver, the the expected memory
should have a sram node for it.

Signed-off-by: Wendy Liang <jliang@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>