phy: add command callback
commit  6384ae5c4946205c1dd1ed73fe9a2c339da6a2ff from
https://github.com/MarvellEmbeddedProcessors/linux-marvell.git

This callback might be useful for complicated phy drivers which require
extra operations, in addition to the existing operations (on/off).

Each new operation requires a new call back, which might be
complicated in some cases. "send command" callback helps to
simplify this, be allowing using one callback for many operations.

The commands type is int, this gives the ability to the users
to define the commands for their use.

Change-Id: Id4470363a37a44c78d7a4e22c9e55254e751f70a
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/37095
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: Omri Itach <omrii@marvell.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>