remoteproc: zynqmp_r5: separate rpu_base and rpu_glbl_base
commit  e4a9f5bbe95ba189140b37c56935e83f0ee91b10 from
https://github.com/Xilinx/linux-xlnx.git

If we have both R5_0 and R5_1 in the system, we need to have two
RPU device nodes in the DTS.

However, both of the remoteproc instance of the two RPUs will access
the same IPI registers, CRL_APB registers (when PM API is used, this
will go away) and the same global RPU global registers.

When sysfs tries to create a link, it will fail because of the same
address. Two possible solutions:
* Add one IPI device node, one RPU register device node, one CRL_APB
  register device node(will go away when PM API is used.)
* separate RPU specific registers and RPU global registers.

This patch uses the 2nd solution.

Signed-off-by: Wendy Liang <jliang@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>