drm: xilinx: dp_sub: Clear the audio reset
commit  1d23f835e441a0271cc64a4ce959d39546fb55ea from
https://github.com/Xilinx/linux-xlnx.git

The audio soft reset register is non-reset flop, and it doesn't get
reset to the default value. So the register needs to be probrammed
explicitly to be specific value. Otherwise, the value varies between
parts, which was resulting in audio not working on specific parts.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>