remoteproc: fix multiple back-to-back error recoveries
This commit comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

The remoteproc core uses a dedicated work item per remote processor
to perform an error recovery of that processor. This work item is
always scheduled upon notification of an error at the moment. The
error recovery process itself is performed when the workqueue gets
scheduled and executes the work function, but an error recovery
needs to be performed only once if there are multiple notifications
while an existing error recovery is in progress. Fix this by adding
a check to make sure the remote processor error recovery work item
is not already running or scheduled.

This fixes an issue with error recovery upon MMU Faults on OMAP
IPU remote processors. An MMU fault on OMAP IPU sends two error
notifications - one a direct interrupt from the MMU, and the second
a mailbox-based crash notification after the remote processor has
collected some backtrace. The mailbox based interrupt mechanism is
used for Attribute MMU (AMMU) faults and other internal exceptions
on the IPU. The backtrace collection on the remote processor is
triggered by the same interrupt which cannot be differentiated
between an MMU fault and an AMMU fault. The remoteproc core changes
in 4.9 kernel around the boot sequences has now caused the second
notification to trigger a secondary error recovery, which was
avoided in previous kernels due to the event detection in the
work-function itself. The newer code sequences changes the timing
w.r.t previous kernels where the recovery process was performed a
bit later due to the asynchronous firmware loading.

Signed-off-by: Suman Anna <s-anna@ti.com>
(cherry picked from commit d35e3721fa1ba7646c4e7cc089e98304a91e98d2)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>