MLK-14454: video: mipi_dsi_northwest: Fix leaking a reset_control
The mipi_dsi_disp_init function takes a reference to a reset_control
during initialization and does not release it.

This becomes a problem on 4.9 because reset_control_get has been split
into exclusive mode(default) and shared(which need to be marked).
Leaking a reference counts looks like attempting to fetch a second
reference to a controller and this WARNs and returns -EBUSY.

Fix by releasing it at the end of the function.

Signed-off-by: Leonard Crestez <leonard.crestez@nxp.com>
[Quanyang:Original patch taken from
https://source.codeaurora.org/external/imx/linux-imx.git imx_4.9.88_2.0.0_ga]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>