crypto: inside-secure: use egress_lock for list and hw ring
commit  7983ba0bc50c278ca93c8fc043ff505f86cbc006 from
https://github.com/MarvellEmbeddedProcessors/linux-marvell.git

The lock must be held for the duration of ring resource
allocation and inserting into the ring linked list.
When the lock is not held for both operations, the HW ring
and the list FIFO get out of sync.

Change-Id: Ic6c09ff491e920d2e35dc66a6b9ae31afd8103d7
Signed-off-by: Ofer Heifetz <oferh@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/37584
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>