ahci: iommu: dts: cp110: add stream ID configuration for SATA unit
commit  010e5352326268dba851528ab271a7257f3f8ec7 from
https://github.com/MarvellEmbeddedProcessors/linux-marvell.git

A stream ID is a unique number that identifies the device
when it requires translation on the SMMU (MMU-500).
The stream IDs for the units in CP-110 are configured in firmware
and are added to Device Tree, to be used by the SMMU driver.

- SATA unit is represented by a single SMMU entry, so added
  a single entry for SATA unit, which in fact represents
  both sata ports.

Change-Id: Ie78e9beb7ed2acbb3d8eab4d72804e093c2e8ef3
Signed-off-by: Omri Itach <omrii@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/38097
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>