drm: i2c: adv7511: Change to CONFIG_DRM_I2C_ADV7511_LEGACY
commit  20216e622e33679fbca5a76dcec7ae8097162ea8 from
https://github.com/Xilinx/linux-xlnx.git

To avoid the conflicts with drm bridge adv7511 driver,
add _LEGACY at the end of the identifier string. This config
depends on !DRM_I2C_ADV7511 so that only one of them can be
enbled at a time. While at it, keep 'depends on OF'.

This is a temporary solution to enable adv7511 based on the encoder
slave. The DRM driver needs to support the atomic modesetting
as well as the drm bridge. When those changes are in, this can be
removed.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>