dma: fsl-qdma: workaround for ERR010812
ERR010812:
Enqueue rejection occurs as a results of the lack of processing by the
consumer of the command descriptors in the status queue. This may be due
to the size of the status queue, i.e. too small, to account for the
delay in reacting to an exceeded queue threshold or other means of
determining a non-empty status queue. While increasing the status queue
size may alleviate the occurrence of enqueue rejections, it is  not a
complete solution. qDMA supports flow control (XOFF) flowing from the
status queue to the command queue(s) producing traffic. This flow
control is initiated when and enter XOFF watermark is triggered as
defined by register SQCCMR. Setting this to the recommended value in the
register description will guarantee that no enqueue rejections will ever
occur.

Signed-off-by: jiaheng.fan <jiaheng.fan@nxp.com>
[Original patch taken from NXP LSDK1709.]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>