MLK-14659 rpmsg: imx: tty: restore support for imx7d
Commit 2a596c353234 ("MLK-13981-2 rpmsg: imx: enable rpmsg virt tty")
enabled support for imx7ulp but it broke support for imx7d since the
currently distributed M4 image for imx7d still uses the old
"rpmsg-openamp-demo-channel" naming.

While we could change the channel name in the M4 image it is better to
maintain backwards compatibility, so add the old naming back to the id
table.

Signed-off-by: Octavian Purdila <octavian.purdila@nxp.com>
[Quanyang:Original patch taken from
https://source.codeaurora.org/external/imx/linux-imx.git imx_4.9.88_2.0.0_ga]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>