MGS-2805 [#imx-526] Enable gcdENABLE_VG by default.
BUILD_OPTION_VIVANTE_ENABLE_VG=1 is always set in
driver_build_sample.sh, so we enable it in the headers by default as
well. The reason behind this is that applications that use
gc_hal_options.h like libGAL do not use the build system to determine if
the macro has been set or not, while the driver passes them on the fly
to the compiler. This fix is required as ioctl() is broken otherwise (
the macro is used inside gcsHAL_INTERFACE and application mistakenly
think that the macro is off while the driver always builds with on).

Signed-off-by: Marius Vlad <marius-cristian.vlad@nxp.com>
[Quanyang:Original patch taken from
https://source.codeaurora.org/external/imx/linux-imx.git imx_4.9.88_2.0.0_ga]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>