drm: xilinx-hdmi-tx: Initial release of xilinx hdmi-tx soft IP driver
commit  bc5656838144a3f09697dd227b5e8212de5124d6 from
https://github.com/Xilinx/linux-xlnx.git

HDMI Tx subsystem is a feature-rich soft IP incorporating all the necessary
logic to properly interface with xilinx PHY layers and provide HDMI encoding
functionality. The subsystem is a hierarchical IP that bundles a collection of
HDMI TX-related IP sub-cores and outputs them as a single IP. The subsystem
takes incoming video and audio streams and transfers them to an HDMI stream. The
stream is then forwarded to the video PHY layer

Driver is being released in staging area and is in experimental state

Signed-off-by: Rohit Consul <rohitco@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>