mtd: nand: arasan_nand: use helper function nand_check_erased_ecc_chunk
When erasing a nand page all the bits are erased to 1s. i.e 0xff but
the ECC of that page can't update in the oob region, hence there will
be ECC errors when reading erased page.

As an example, when we use UBIFS on NAND, UBIFS assumes that empty space
is really empty and the below layer deals with bitflips in empty pages.
So, when we enable HW ECC, and when the erased command issued, then
the page will be erased with out ECC update. and while reading arasan
controller detects ECC error and corrects the erased page data (not 0xffs).
With this data when UBIFS is checking for ffs in erased page, it won't
find proper ffs causing UBIFS errors.

So to avoid this, driver should correct bitflips in the buffer and pass
that to upper layers. This can be solved by the API
nand_check_erased_ecc_chunk.

Signed-off-by: Naga Sureshkumar Relli <nagasure@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit 6216d9d254bebb803665a00d143306de77e36dee)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>