usb: xhci: Add toggle cycle bit for the last seg trb when cached ring is used
At present USB 2.0 cameras fail to play when run for 2nd time or greater.
This issue occurs only when previous endpoint cached ring is used instead
of allocating a new endpoint ring. The root cause of the problem narrows
down to the Toggle bit, which is not getting set on the last trb of the
cached ring last segment. Since the toggle bit is not set for the last trb
of the last seg, controller generates RING OVERRUN error and stops
processing the further TRB's , resulting in no video playback for 2nd run.
This patch solves that problem by adding the TOGGLE bit for the last trb
of the last segment of the previous cached ring.

Signed-off-by: Anurag Kumar Vulisha <anuragku@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit e7f7d5614e4a873aae84903ef13d1a0b3ecf8951)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>