remoteproc: add a rproc_set_firmware() API
This commit comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

A new API, rproc_set_firmware() is added to allow the remoteproc client
drivers to be able to configure a custom firmware name that is different
from the default name registered by a remoteproc platform driver. This
function is being introduced to provide a kernel-level equivalent of
the current sysfs interface, so as to allow these kernel-drivers to
choose different firmwares based on their functional feature it is
implementing using the remote processor.

Signed-off-by: Suman Anna <s-anna@ti.com>
(cherry picked from commit b1fe0b7510b64ca520ceda3293961533f4b8ce67)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>