remoteproc/pruss: update pruss_get() to retrieve a PRUSS id
This commit comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

Update the pruss_get() function to take in an additional integer
pointer argument in which the PRUSS instance id is filled in and
provided back to the callers. This allows the drivers to add some
instance-specific logic/customization in their code, as the PRUSS
handle is not useful to build this logic.

Signed-off-by: Suman Anna <s-anna@ti.com>
(cherry picked from commit 91157e8787c1b02214b4e468b596cfc8dcc4dcfc)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>