MLK-15080 PCI: imx: pcie ep can't be probed properly
iMX7D Sabre SD board implement the GPIO expander
connected to a peripheral bus.
Probe deferral would be triggered when try to request
the expanded GPIO at the first time.
pcie ep can't be probed properly at the second probe,
because of the duplicated registration of the sysfs.

Change the registeration point of the sysfs to fix
this issue.

Signed-off-by: Richard Zhu <hongxing.zhu@nxp.com>
[Quanyang:Original patch taken from
https://source.codeaurora.org/external/imx/linux-imx.git imx_4.9.88_2.0.0_ga]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>