net: mvpp2: initialize the XLG MAC when using a port
commit  a853adbaf2fd9ede70b12c63a137ffb3a9569d7a from
https://github.com/MarvellEmbeddedProcessors/linux-marvell.git

This adds a routine to initialize the XLG MAC at the port level when
using a port and the XAUI/10GKR interface mode. This wasn't done until
this commit, and the mvpp2 driver was relying on the bootloader/firmware
initialization. This doesn't mean everything is configured in the mvpp2
driver now, but it helps reducing the gap.

Change-Id: If21a9192fcb465ee575fa161d0c4eff9f826e081
Signed-off-by: Antoine Tenart <antoine.tenart@free-electrons.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/42828
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>