net: ethernet: xilinx: Fix ping not working after ifdown and ifup
With the existing driver if user does a ifdown and ifup
of the ethernet interface data integrity is not working.
We are stoping the ndev in the close/down but not starting
the ndev in the open/up.

This patch fixes this issue by starting the ndev in the open/up().

Signed-off-by: Kedareswara rao Appana <appanad@xilinx.com>
Acked-by: Harini Katakam <harinik@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit 8b9398cefef530d063f250b92944fb2dd1349411)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>