arm64: zynqmp: Change Kconfig dependency for cortexa53 edac
The a53 edac driver never got accepted by mainline and it will never be.
The code that is creating issue is:

	get_online_cpus();
	for_each_online_cpu(cpu) {
	/* Check CPU L1 error */
	smp_call_function_single(cpu, parse_cpumerrsr, NULL, 0);
	}

We check to find out the online cpus and for each available cpu we
schedule a L1 check. Rarely for corner cases, by the time we schedule
a L1 check in a CPU, that CPU goes offline. But the problem is when
it happens it is a kernel crash. It is a severe problem.

So with CPU_IDLE enabled, A53 Edac won't work. So disable this a53
edac by default if CPU_IDLE is enabled via Kconfig dependency.
And if CPI_IDLE is disabled then by default a53 edac will be enabled
with Kconfig default option set to y.

Signed-off-by: Naga Sureshkumar Relli <nagasure@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit 3fe9aa9eca889f17a1b419cc787ac03441f2f2f4)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>