MLK-15004: ASoC: fsl: Refactor DMA workaround
Commit 2f756e7aa8840 ("ASoC: fsl_esai: esai workaround for imx8qxp
Rev1") introduced a workaround for ESAI.

Because the same workaround needs to be done for SPDIF, we refactor
GPT handling in order to avoid code duplication.

Notice that we isolate code related to workaround into
fsl_dma_workaround so that only few lines of code from ESAI/SPDIF
are modified. Thus when the hardware issue will be fixed there will
be very little things to revert in our modules.

Reviewed-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>
[Quanyang:Original patch taken from
https://source.codeaurora.org/external/imx/linux-imx.git imx_4.9.88_2.0.0_ga]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>