sdhci: add support for switching to UHS-I modes on ZynqMP Platform
This patch adds support for switching to UHS-I modes on ZynqMP platform.
There is one issue with the ZynqMP ZCU102 where it is not able to power
cycle the SD card. Due to this once the card is initialized in UHS-I
mode it cannot be reinitialized in UHS-I mode again. So to solve this
issue there are some of the checks are made in place to ensure that the
card operates in UHS-I mode even on reinitialization. This will enable
the SD host controller to run SD at 200MHz in linux if it was running at
200MHz in U-boot.

Signed-off-by: Manish Narani <mnarani@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit dcbfca5021bbd5160f0f2e8f42b7bf6aca2d02bd)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>