soc: ti: ti_sci_pm_domains: Populate name for genpd
This commit comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

Commit b6a1d093f96b ("PM / Domains: Extend generic power domain
debugfs") now creates a debugfs directory for each genpd based on the
name of the genpd. Currently no name is given to the genpd created by
ti_sci_pm_domains driver so because of this we see a NULL pointer
dereferences when it is accessed on boot when the debugfs entry creation
is attempted.

Give the genpd a name before registering it to avoid this.

Fixes: 52835d59fc6c ("soc: ti: Add ti_sci_pm_domains driver")
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Signed-off-by: Santosh Shilimkar <ssantosh@kernel.org>
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
(cherry picked from commit 4dd6a9973b8aaffac4bf37c5bb70e8eae5a7afb4)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>